mpegtsenc: Add support for output of SCTE-35 packets
Introduce the ability to pass through SCTE-35 packets when creating
MPEG transport streams.  Note that this patch makes no effort to
change the PTS values in the SCTE-35 packets, and thus only works
when not reclocking the stream.  A subsequent patch includes a BSF
to recompute the PTS values.

Remuxing Example usage (assumes video on 0:0, audio on 0:1, scte on 0:2):

./ffmpeg -copyts -vsync 2 -i scte35.ts -map 0:0 -map 0:1 -map 0:2 -codec:a copy -codec:v copy -codec:2 copy -f mpegts -muxrate 14M output.ts