Add new BSF to recompute PTS timestamps in SCTE-35
Because SCTE-35 messages are TS sections rather than PES packets
we cannot rely on the standard mpegts mux to adjust the timestamps
as needed when reclocking the stream.  Introduce a BSF which can
recompute the PTS timestamps, taking advantage of the SCTE-35
pts_adjust field to accommodate cases where the output timestamps
don't match the input.

The mpegtsmux will automatically add the BSF into the pipeline,
so there is no need for the user to specify a "-bsf scte35ptsadjust".

Example usage:

./ffmpeg -i scte35.ts -map 0 -codec:a copy -codec:v copy -f mpegts -muxrate 14M output.ts