Only insert interlacing filter if video is actually interlaced
The addition of the "-interlace" command line argument was causing
the interlace filter to always be inserted into the pipeline.
Because the filter itself isn't smart enough to pass through
progressive video, we are just going to avoid inserting the filter
into the pipeline if the AVFilterLinks indicate that the video
is progressive.