lted bug #16: Fix incorrect selection of audio track
Our general approach is to output audio streams based on the order
they appear in the PMT.  We do this if LTED is configured to output
all streams, but if configured to only output one stream it falls
back to ffmpeg's default selection algorithm rather than picking the
first stream.  This can lead to unexpected behavior.

Rework the logic such that if we're not told to output all streams
we choose the first stream in the PMT, rather than the "best"
stream according to ffmpeg's legacy logic.