mpegtsenc: Don't periodically announce PCR on SCTE-35 streams
Changes were made between in the last two years to periodically
send PCR-only packets on all PIDs, but for cases where the stream
may send packets very infrequently (like SCTE-35), this results in
extra TR101290 errors because it fails the PCR interval test.

I am not quite sure what the "right" fix should be for this, but
for now just disable all periodica sending of PCR-only packets on
SCTE-35 streams.