Add hack to skip deinterlacing/interlacing in certain cases
In cases where the resolution and framerate are unchanged, and
both the input and output are interlaced, don't deinterlace and
re-interlace the video.

This is primarly to avoid a quality and performance hit in the
case where the output is forced to some video mode and the stream
arriving is already in that same mode.