vf_pad: make recent change to ensure height divisible configurable
In a recent commit I hard-coded a hack to ensure the output height
was always divisible by 8, in order to work properly if swscale was
further down the pipeline.  However this breaks the case where we
pad 480->486 in order to satisfy the decklink output.

Add a new variable to the filter, which allows the caller to
specify that the height must be divisible by some number.