smpte2038: Port support for passthrough of SMPTE 2038 from lted1 branch
Forward-port the SMPTE 2038 passthrough support from the lted1
branch.  Most of the changes are pretty basic and related to adding
a new codec type, setting the PMT registration descriptor
properly, and making sure the stream is treated as sparse.  But
there is one change which is worth mentioning:

Between our lted1 branch and master there was a fix to the mpegts
demux which resulted in the stream_id being properly sent through
the pipeline.  This exposed a second bug in the mux where any
stream of type data which had a stream_id of 0xbd would have its
PTS values removed.  That change should only be applicable if
it's a KLVA stream, so change the logic to prevent other stream
types from having their PTS values stripped in the output.