Fix possible divide-by-zero condition during pre-roll
When prerolling we ensure we don't queue up more than 500ms worth
of audio to prevent overflowing the hardware FIFO.  But if the
encoder sends a large burst of video without any audio at startup,
we might have the audio callback fire before any audio is queued.
While this should have been a no-op since the audio queue itself
is empty, before iterating the queue we compute how many packets
to keep based on the size of the audio packets.

If we haven't received any audio packets yet, the size is zero and
thus we have a divide-by-zero condition.

Augment the if() statement to skip the entire process if we haven't
received any audio yet.