decklink: Fix VANC reference counting
In order to schedule the same frame multiple times, the reference
counting code for the VANC section of the decklink video frame must
be correct.  It turns out I actually implemented this in the version
merged upstream in response to comments from the decklink maintainer,
but never backported it to LTN's tree.