Refactor audio/video insertion on slow streams
Rework the code to move responsibility for the injection of
frames into its own function, and do both audio and video (i.e.
reschedule the same video frame multiple times as needed).

This is specifically targeted at streams which arrive with *almost*
correct timing but are slightly slow (i.e. off by a few frames per
hour).  This would normally cause the FIFOs to eventually drain out
and result in audio loss and/or video freezes.  Gracefully recover
from that condition when we hit a low water mark, bringing it back
to the preroll level.