Move LTN log initialization to main program
In the original implementation the LTN logging facility was setup
when initializing the decklink output.  This wasn't a problem since
all the log events came from the decklink module.  Now that we're
using it in other parts of the application we need to set it up earlier
in the process (since some messages may be generated before the
output is initalized).

This also will help in cases where the decklink output isn't used
at all (e.g. for the transcoder/thumbnailer projects).