Have NDI library be loaded via dlopen()
Dynamically load the NDI library so that if NDI is not needed we don't
require all of its dependencies to be installed on the platform
(i.e. avahi-client, etc).

Of course if you want to use LTED with NDI you need to have the NDI
library as well as any dependencies it has installed.

This is essentially a version of the patch sent to the ffmpeg-devel
mailing list which never got merged.

djh - Fix merge conflicts