Add a new filter to repack CEA-708 arrays
In some cases an encoder may send 708 tuples which include 708
padding between valid 708 packets.  This is technically legal in
the CEA-708 standard, but discouraged in the CEA Implementation
Guidance (CEA 10-B: 708-B Implementation Guidelines, Sec 26.5)

Reuse the CC repacking code that we built for video output conversion
to create a generic filter which can be applied even when framerate
conversion isn't being done.  This strips out any padding and
reconstructs valid arrays of tuples where the padding is only at
the end.