Fix issue where latency may slip after decklink output reset
When resetting the output, the first_pts field should be set back
to zero, rather than the frame that prompted the reset.  Otherwise
in instances where there is a stall followed by a burst containing
the old data the first_pts will be pointing to a very old frame
even though we discarded the subsequent 90 frames of video.

Also, don't announce the frame that caused the reset to the actual
hardware.