Rework insertion code to use same basic logic as drop frame code
A recent change was made to drop a frame per minute if the current
number of frames bufferred exceeded the target preroll level by
a certain amount.  Go back and implement the same change if we are
below the target preroll level.

This means we no longer wait until we hit the low watermark and then
insert four frames, but now proactively insert a frame as soon as
we're below the target preroll level.  As a result the change is
more subtle (e.g. one frame inserted per hour instead of a four frame
jump after four hours), and it avoids some edge cases where we
have dipped below the low watermark for audio but haven't yet reached
that point for video (and hence there is a window of audio stuttering
until the logic kicks in).

Also, because the decision is based off a statistical average over
the last minute rather than the buffer level at the latest frame
queued, you are less likely to have latency suddenly pushed out due
to very short periods of packet loss (e.g. 200ms).