Change logic for determining preroll value
The problem attempting to be solved here relates to cases where there
could be hundreds of milliseconds of audio packed into a single PES
packet, which means we need more latency in the pipeline to prevent
audio stutters.  We originally solved that by calculating the
difference between PTS values in adjacent PES packets (i.e. a PES
with many AAC audio packets would have a large PTS delta).  However
this exposed a bug where if there was a network dropout during the
probing window the PTS delta between two packets could include the
entire period of the outage (i.e. a 300ms network dropout would
result in a PTS delta of 312ms).

Change our approach to rather look at the different packets with
a given PES and use that to compute the total PTS time that packet
represents.  This accomplishes the same result but isn't dependent
on contiguous PES packets being delivered without discontinuity.