Fix stall on exit with certain content types
The changes to stash the pipeline statistics in AVFrames/AVPackets
created an issue where ffmpeg would fail to exit when the input
hit end-of-file.  This was because we were inserting the stats into
side data even for "flush" AVPackets (packets with no payload used to
flush the decoder), and items further downstream were looking for
AVPackets that either had a payload or side-date to determine whether
it was time to exit.