mpegts: Fix incorrect insertion time for pipeline stats
The code which inserts the AVFORMAT_INPUT_TIME timestamp for the
pipeline stats grabs the current time once the PES is full received.
The correct behavior is to use the time the PES started instead.  So
stash the timestamp at the start of the PES so it can be inserted once
the PES is completely assembled.

This was causing latency timestamps to be off by about the duration
of a video frame (16ms for 59fps, 33ms for 29fps).