Add config option to enable/disable "live" behavior
We recently introduced changes to more closely track the preroll
level (to provide more deterministic latency).  However this should
only be used with streams arriving in realtime, since with canned
file content the expectation is the output provides the rate control.

Add a flag to enable this behavior, so that we don't drop one frame
per minute when doing file playout.