Expose LTED state machine to controller
Report back to the controller as the LTED process transitions
through various states of operation.  This allows us the controller
to make better decisions about whether to restart the process as
well as making it easier to interpret various statistics (e.g.
some stats are not valid during the probing phase).