Expose counter of UDP packets dropped by operating system
For cases where the UDP receiving thread isn't able to keep up
with the operating system, packets will queue on the kernel
thread until they are dropped.  Don't just keep track of the
queue level, but also track the count of packets dropped.  This
makes it more obvious when there are dropped packets even when
the reported queue level looks ok (which can happen with very large
bursts that are very short).