Add new pipeline stat between mux output and decklink input
Before we were tracking when the write to the mux occurred and
after the frame was played out.  However without any intermediate
points we couldn't tell how much time the frame was spending on
the output reordering mux.

Add a data point to record when the avformat write_packet() callback
is fired.