First cut at making resampling work within AC4 decoder
Despite the base frame containing 1536 samples, the codec supports
the ability to scale that to some arbitrary output sample count.

See TS 103 190-1 - V1.3.1, Sec 6.2.15: Sampling rate converter

In the sample I'm testing with, the base frame contains 1536 samples,
which would represent 32ms at 48Khz, but the actual time intervals
for that block is 33.3ms, so we have to stretch the number of samples
delivered (frame_rate_index is 3, so we're supposed to resample at
a ratio of 25025/24000).  See Table 83 for the resampling ratios
based on the frame_rate_index).