Exit if encountering massive PTS jump in FPS filter
If one input stream stops and another starts and video is on the
same PID, the vf_fps filter will try to create all the intermediate
frames between the last frame in the old stream to the first frame
in the new stream.  This can result in trying to create thousands
of frames.  This is because stuff like ffmpeg's built in
non-monotonic PTS detection is in the output phase of processing,
which is after the AVFrames have been routed through the filtergraph.

While there should probably be code further up the input chain to
detect stream switches, putting the logic inside the FPS filter
addresses the immediate issue while only being in the code path
if output format conversion is enabled (thereby minimizing risk
of causing breakage).  This logic can be disabled by setting
the environment variable "LTN_SKIP_RESTART_ON_PTSJUMP".