Working state for AC-4 support
This is a whole pile of stuff in my working directory representing
the current state of development.  It includes:

Paul Mahol's original patches
ixes I made to the MPEG-TS mux and MPEG-TS demux
A raw muxer for AC-4 (used by the MPEG-TS mux)
A hack to the DASH decoder to make it work with our ATSC3 streams
  (need to figure out the "right" fix)
A fix to Paul's AC-4 raw demuxer causing probe to bail out early

This obviously needs to be broken out into a full patch series, but
at least get it into source control as-is so Graeme can use it.

Note, this patch was backported from branch graeme-ac4, with minor
tweaks to support the API changes between 4.4 and 4.5