Send SEI unregistered messages out the NDI interface as metadata
Deliver SEI unregistered messages to downstream NDI receivers, and
in particular the messages used to compute glass-to-glass latency
when using the LTN encoder.

Note this patch has some extra bits to use the NDIlib_video_frame_v2_t
API and struct, which is newer and allows for delivery of per-frame
metadata.