Port ltnlog facility to latest tree
The libavformat/ltnlog facility is used for reporting back
statistics to the controller script which runs the ffmpeg process.
These include things like input stream properties as well as running
stats for various FIFOs, counters, etc.  These stats are both shown
in the user interface as well as used to make decisions on the health
of the ffmpeg process (e.g. force a restart if the decklink output
frames per second drops to zero).

Subsequent patches will show calls to ltnlog_stat() sprinkled
throughout the codebase at various key locations where we need
visibility into the operation and running state of the process.