Support LTN's non-standard naming for decklink devices
In the upstream tree you can either refer to a decklink device based
on it's model name (e.g. "Decklink Duo (3)") or by it's PCI ID and
slot.  Neither of these are particularly well suited for our use
case.

Re-implement the naming convention we've been using for several
years, where we refer to devices as "decklink[0-8]"