decklink: Add support for output of static HDR metadata
This patch is a WIP and needs more cleanup, but basically works
well enough to play out the "Gemini Man" sample we use internally
for testing.

Note there are unrelated issues that need to be investigated with
this clip related to audio and intermittent stuttering, but that
is unrelated to the HDR support.