Return zero if values are unknown, and only E_INVALIDARG if not unknown value
Returning E_INVALIDARG causes the frame submission to fail, so only
use it if we get an actual invalid parameter requested.  Otherwise
return zero for values if unknown.

For example, a clip may have mastering info but lightlevel SEI
metadata.  In such a case we still want the clip to play out, but
with the max_content_light_level and max_pic_average_light_level
set to zero.