Only respond to colorspace callback if supported by the card
The SDK issues the callback to get the colorspace even if the card
doesn't support setting the colorspace.  For cards that are older
and don't support Rec2020, setting the colorspace to 2020 will cause
the ScheduleVideoFrame() call to fail.

Only provide the colorspace if the card supports such functionality.

This change was done in consultation with Blackmagic Support.