bsf: Add new bitstream filter to set SCTE-35 pts_adjustment when reclocking
Because SCTE-35 messages are represented in TS streams as sections
rather than PES packets, we cannot rely on ffmpeg's standard
mechanisms to adjust PTS values if reclocking the stream.
This filter will leverage the SCTE-35 pts_adjust field to
compensate for any change in the PTS values in the stream.

See SCTE-35 2019 Sec 9.6 for information about the use of
the pts_adjust field.

This filter also tweaks the mpegtsenc mux to automatically
add it so the user doesn't have to include it manually.

Thanks to Andreas Rheinhardt and James Almer for providing
feedback/suggestions on improving the patch.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>