Break out counts for individual 708 service numbers
Previously we were just counting the total service blocks, but since
many services blocks will be for service block 0 this may be a
bit misleading with regard to whether there are valid captions
present on the other services.

Change the struct so we can track the total count for each service,
and reflect that in the report.