Make logging of byte buffers atomic
The debug code which dumps out groups of bytes was previously
calling av_log() in a loop.  However this can result in log
lines being interrupted by av_log() calls from other threads
(making the logs difficult to read).

Switch to using av_bprintf() to build up the message to be logged,
and then have a single call to av_log() which is atomic.