Move the initialization of data/subtitle streams until after their timebase is set
We need to set the timebase of any data streams to match the video, so that
we can insert the data on the right video frame.  This requires us
to setup the video, which includes setting the timebase on the
video stream, and then setting the timebase for any data streams.

Because any automatically added BSF filters will set their timebase
during initialization, we need to set the stream's timebase before
we do initialization.  Otherwise, we end up doing timebase conversion
twice (both before write_packet as well as after going through the
BSF, which results in garbage PTS values since it's using the 90Khz
timebase established when the bsf was initialized).

Convert to doing two passes over the stream list.  The first deals with
audio/video and estabilishes the video's timebase.  Then set any
data streams to the video's timebase.  Then loop over the stream list
a second time and initialize the data streams.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>