decklink: Add support for output of multiple audio channels
Add the ability to put out multiple audio streams, including a
mix of raw PCM streams audio and bitstream audio.  Because packets
for each stream are announced separately but we need to interleave
them all before announcing the audio to the hardware, we use an
intermediate queue, and as we get closer to the playback time we
only then schedule the audio buffers to the hardware.

This support also eliminates the constraint that an audio stream
must have 2, 8, or 16 channels.  This is particularly important
because you couldn't output 5.1 audio since it didn't handle
six channels.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>