hevc: Fix propagation of interlaced status, including TFF/BFF
Look in more detail at the pic_timing info to support showing the
correct field order information both in the probing phase (i.e.
as part of the AVCodecContext), as well as as part of the AVFrame
fields that describe the frame structure.

The value from the SEI "picture_struct" was being blindly copied
into both s->picture_structure as well as s->field_order (even
though one of of type AVFieldOrder and the other is
AVPictureStructure).  In some cases this resulted in the wrong info
being populated, and in other cases the s->picture_structure would
get populated with AV_PICTURE_STRUCTURE_UNKNOWN because the
pic_struct value wasn't one of the six values recognized
in the SEI callback (this includes progressive video).

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>