Add filter to construct frames from interlaced HEVC fields
Because the HEVC decoder doesn't know how to construct AVFrames
composed of field pictures, we use a filter to take the AVFrames
containing individual fields and create full frames.

This is a hack.  The codec should really be augmented to construct
proper AVFrames, but that is a very invasive change.

Note, this filter can always be present in the filter chain.  If
we're not receiving interlaced video at 1920x540 or 720x240, we
just pass through the frames unmodified.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>