ltndecoder: Add debug option to dump filtergraph to stderr
Add a command-line option which will log the negotiated filtergraph
to the logfile.  This makes it easier to see the final pipeline.

Note this patch doesn't implement the full support in LTED v1 which
put out a Graphviz document to a file.  This is essentially just a
port of a patch from Liman Wang which never got merged upstream.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>