v210enc: Add YUV420P as a supported input format
The current V210 filter would only accept 422P and 422P10, and thus
in the most common use case we encounter (YUV420 from a decoder),
the pipeline would get an swscale instance inserted to convert the
colorspace.

However, while the swscale filter performs fine when scaling video
and changing the format, it is poorly optimized for the case where
the width/height of the video, as well as the bit depth is unchanged.
The net effect is it's running the full scaling algorithm against
a line that doesn't require any scaling or bit depth conversion,
where a simple memcpy() could achieve the same result.

While we could improve swscale for what they clearly saw as an
edge case, we can do even better and skip the scaler entirely by
augmenting the V210 encoder to directly work with 420P.  In this
patch, we simply reuse the same source chroma line for each pair
of lines in the output.  With this change, we can go straight from
the output of the decoder to the v210 encoder, without having to
touch the pixels at all in between.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>