Properly set interlaced/tff flags on processed frames
The phase filter changes the effective setting of tff/bff but
it wasn't being reflected on AVFrames coming out of the filter.

Change the filter to properly set the flags based on how the
frames were processed.

Note, we also throw away the first field, because until the second
frame arrives the frame gets passed through as progressive, which
screws up downstream negotiation of the output (which is derived
based on the first frame of video coming out of the filtergraph).

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>