v210enc: Refactor side data passthrough
As the number of side-data items we pass through increases, refactor
the values into a table in the same manner as done in decode.c.

In the future it might make sense to make this a shared function
which can be reused by other encoders.  But for now keep it local.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>