decklink: Add support for output of HDR metadata
Add HDR support to the decklink output for cards that support such
functionality.  This includes setting the EOTF, the colorspace,
the mastering info, and the content light level info.  Both the
Payload Identification HANC data as well as the SMPTE ST 2108-1
VANC data are being set.

Tested with in-house content as well as samples from 4kmedia.org.
Testing was done with the Decklink 8K Pro and the Duo2 with 12.5.1
firmware, as well as with the Duo2 with 10.11.2 (before it supported
HDR) to ensure there are no regressions.

Signed-off-by: Devin Heitmueller <dheitmueller@ltnglobal.com>