Fix memory corruption introduced by pipeline stats patch
There was some refactoring of the support for pipeline statistics
to support changes to the way the encoder/decoder codecs are called,
and I didn't put the call in the right place, which would sometimes
result in either a NULL pointer dereference, or referencing a frame
which hadn't been populated yet.  Both would cause segfaults.