Don't output malformed ST2010 packets if too large
The libklvanc library doesn't support for fragementing SCTE-104
packets across multiple ST2010 packets in the SDI domain.  For now
don't feed the parser if the packet is too large, as the current
implementation will output a malformed frame.

Once support for creating multiple ST2010 packets is supported,
this check can be removed.