Fix regression in decoder pipeline latency caused by 52b67f34bc
The check to make sure the pkt->size is nonzero broke the mpegts
demuxer, where I was setting the input time before I actually set
the payload on the packet.

Move the code which sets the input time a few lines down so it's
after the addition of the payload.