Fix audio distortion when playing TS with incorrect PTS values
We encountered an encoder which produced streams where the PTS
values were off by only a few samples, and averaged out just
fine over time.  In this case we are better off concatenating the
uncompressed blocks of audio samples than introducing arbitrary
gaps and/or overwriting old samples.

The problem manifests as "static", given it occurs rougly once
every 1400 samples (out of 48000/second).

Of course we don't want to *always* concatenate samples, as that
would result in A/V sync slips if there are entire audio blocks
lost.  So we only concatenate if the PTS values are off by just
a few relative to the expected value.

Yes, this is a terrible hack, but it's to work properly with
a commercial encoder product.