Don't attempt to create an input stream list if we can't find PAT/PMT
The TS demux has some functionality to attempt to construct a list
of streams based on PES packets arriving, presumably as an attempt
to play out stream with non-existent PAT/PMT.

This should never happen with a live TS, and certainly not with
broadcast quality equipment.  Further, the "fuzzy" attempt to
construct an input list can be error-prone, so we are better off
bailing out right away than proceeding with bad data.