Fix deadlock if preroll is more than 1.0 seconds
There is a built in cap of 60 buffers, but the code that caps
the buffer count is after the setting of the preroll.  The
effect of that is that you can end up with a preroll that
exceeds the number of buffers actually allocated, which will
cause a deadlock (i.e. the routine to queue new buffers will
block, but the stream won't have started outputting frames yet).

Change the logic such that we set the preroll first and cap that,
and *then* set the buffers to 2x preroll.

Note, also move the line which reports back the preroll to the
controller to be after all the calculations are done, so that
it reflects the final value.