Fix indefinite stall on decklink output reset
The upstream maintainer recommended using AV_NOPTS_VALUE for
the first_pts member, rather than zero.  I agreed with the
recommendation but missed a case when backporting the LTN patches
for resetting the output.

The effect would be that if we pulled a reset due to falling
behind schedule the output would stall forever because
the first_pts would be set back to zero, and thus not be changed
to the PTS of the next video frame received.

Set the first_pts member to AV_NOPTS_VALUE when resetting the
output.