Perform colorspace conversion for 601/709 if doing output conversion
If doing scaling/framerate conversion, the application should also
be converting from 709 to 601 if output is SD or 601 to 709 if output
is HD.  With newer decklink firmwares they've made it explicit and
the scheduling of frames will now fail if the colorspace isn't correct
for the SD output.

Have the application add the colorspace filter to the pipeline to do
the necessary conversion as needed.