Bail out if one of the VANC line arguments specified is -1
We allow the user to disable the output of a specific VANC type
by setting the line number to -1, but we missed adding the checks
which cause the routine to bail out.  The effect is it would attempt
to call the video frame VANC insertion function with an invalid line
number which would spew errors to the console for every frame.