Set lower bound for preroll to 3 frames
Even if the user specifies an arbitrarily low preroll value, don't
let the actual calculated preroll go below 3 frames.  This is
because we with 1080p24/1080p25 we would end up with only two
frames of preroll when the user set the preroll to 0.1 (since the
calculation uses integer arithmetic).

Having a preroll of two frames results in continuously reporting
that the low watermark has been hit, but because the preroll is
already two frames you end up "Advancing 0 frames" (since the
insertion logic was intended to bring the buffer count back to the
target preroll level).