Fix case where state machine was being updated continuously
Due to a mistake when backporting the patches, I put the routine
which notifies the controller that the decoder is fully running
into the main loop.  This caused it to continuously update the
controller.  Because we clear the watchdog on state machine
changes, this meant the decoder wasn't being restarted if the
output stopped but the input was still arriving (as may happen
for example on a PTS jump backwards).

Remove the call, so we only notify the decoder once that it's
reached a fully running state.