Support I420 in ffmpeg NDI output module
In most cases we'll receive TS streams in H.264/HEvC, where the
most common format will be is I420.  Because the conversion to
UYVY using swscale is so expensive, we're much better off just
feeding the NDI output module with I420 frames and letting the
SDK to the conversion to UYVY.

Note this requires a bit of a hack because the NDI library expects
all three planes to be contiguous in memory (i.e. a single buffer
containing the Y, followed by U, followed by V), and AVFrames
typically have the three planes in separate buffers.  So we
combine the planes into a temporary buffer before handing off to
the SDK.