Make audio preroll configurable
While 50ms is generally enough headroom to ensure we have enough
time to interleave the audio from all the different streams, there
are cases where genlock is present that result in clock skews which
put us too close to the deadline.

Make the audio preroll configurable via a command-line option.