Provide CC data in UTF-8
The libzvbi library provides the CC characters as Unicode codepoints,
so we shouldn't simply treat them as ASCII characters.

Properly convert them to UTF-8, so the receiving application can
decode non-ASCII characters.

Note that the CC decoder itself doesn't appear to properly support
Extended character sets for languages such as Spanish, but that is
a separate problem.