Port stream selection logic from LTED tree
We want to pick the first video/audio streams from the first
program, so port that logic over from LTED.  Without this, on an
MPTS it could end up picking the video from one program and the
audio from another.  This also ensures that if all audio streams
are to be decoded, that it only includes those streams within the
first program.