Move HEVC field merging and deinterlacing to output filter section
Later versions of ffmpeg refactored the filtergraph negotiation,
including completely removing deinterlacing.  When we added it back
in along with the HEVC field merging we put it in the input filter
graph.  But this causes negotiation failures because the output
filter graph construction uses the original values.

Move the two filters into the output graph section to simplify the
logic.